requirements.txt
streamlit==1.45.1
pandas==2.3.0
pytest==8.4.0
fastapi==0.115.12
pydantic==2.11.5
uvicorn==0.34.3
mysql-connector-python==9.3.0
requests==2.32.4
